+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## [Unreleased] - ReleaseDate
 ### Changed
-- Dropped the `users` crate and fixed erroneous user lookups
-  ([Issue](https://github.com/fubarnetes/rctl/pull/5),
-  [PR](https://github.com/fubarnetes/rctl/pull/5))
+- Updated the Users crate (#1, #7)
 
 ## [0.0.5] - 2018-12-20
 ### Added